pyproject.toml
[build-system]
requires = [ "poetry>=1.0.0",]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
py_version = 37

[tool.black]
line-length = 79
target-version = [ "py37",]

[tool.poetry]
name = "ERPLibre"
version = "1.2.1"
description = "Easy way to configure Odoo community"
license = "AGPL-3.0-or-later"
authors = [ "Mathieu Benoit <mathben@technolibre.ca>",]

[tool.poetry.dependencies]
python = "^3.7.12"
Asterisk = "^0.0.8"
Babel = "2.9.1"
BeautifulSoup4 = "^4.10.0"
GitPython = "3.1.12"
Jinja2 = "2.11.3"
Mako = "1.0.4"
MarkupSafe = "0.23"
Pillow = "9.0.0"
PyJWT = "1.7.1"
PyPDF2 = "1.26.0"
PyYAML = "4.2b4"
SOAPpy = "^0.12.22"
Sphinx = ">=2.4.0"
Voicent-Python = "^1.0"
Werkzeug = "0.11.15"
XlsxWriter = "0.9.3"
agithub = "^2.2.2"
altair = "^4.2.0"
argcomplete = "^2.0.0"
astor = "^0.8.1"
bokeh = "1.1.0"
boto3 = "1.9.102"
cachetools = ">=2.0.1"
cerberus = "^1.3.4"
chardet = "3.0.4"
click = "^8.0.3"
cloudflare = "^2.8.15"
code_writer = "^1.1.1"
colorama = "^0.4.4"
cryptography = "3.4.8"
css-html-prettify = "^2.5.5"
cython = "^0.29.26"
ddt = "1.2.0"
decorator = "4.0.10"
docutils = "0.16"
ebaysdk = "2.1.5"
email_validator = "^1.1.3"
emoji = "^1.6.3"
escpos = "^1.9"
extract-msg = "^0.30.0"
extract_msg = "^0.30.0"
factur-x = "^2.3"
feedparser = "5.2.1"
flake8 = "^4.0.1"
formio-data = "^0.3.14"
freezegun = "0.3.11"
geojson = "2.4.1"
gevent = "1.3.4"
giturlparse = "^0.10.0"
greenlet = "0.4.13"
html2text = "2016.9.19"
html5print = "^0.1.2"
invoice2data = "^0.3.6"
iscompatible = "^0.1.1"
isort = "^5.10.1"
jira = "2.0.0"
js2py = "^0.71"
keystoneauth1 = "3.14.0"
lasso = "^0.0.5"
libsass = "0.12.3"
lxml = "4.7.1"
matplotlib = ">=2.0.0"
mock = "2.0.0"
mpld3 = "0.3"
mysqlclient = "^2.1.0"
num2words = "0.5.6"
numpy = "1.21.1"
oauthlib = "2.1.0"
"oca.decorators" = "^0.0.1"
ofxparse = "0.16"
openpyxl = "^3.0.9"
openupgradelib = "^3.3.1"
pandas = "1.2.0"
paramiko = "^2.9.2"
passlib = "1.6.5"
pdf2image = "^1.16.0"
pdfminer = "^20191125"
pexpect = "^4.8.0"
phonenumbers = "^8.12.41"
plotly = "4.1.0"
premailer = "^3.10.0"
psutil = "5.6.6"
psycopg2 = "2.7.3.1"
py-Asterisk = "^0.5.18"
"py3o.formats" = "^0.3"
"py3o.template" = "^0.10.0"
pycountry = "^22.1.10"
pycrypto = "^2.6.1"
pydot = "1.2.3"
pygount = "<1.2.0"
pyjsparser = "^2.7.1"
pyldap = "2.4.28"
pymssql = "^2.2.3"
pymysql = "^1.0.2"
pyotp = "^2.6.0"
pyparsing = "2.1.10"
pyproj = "3.2.1"
pyserial = "3.1.1"
pysftp = "^0.2.9"
python-chess = "<0.24"
python-csv = "^0.0.13"
python-dateutil = "2.7.3"
python-git = "^2018.2.1"
python-json-logger = "0.1.5"
python-keystoneclient = "3.22.0"
python-slugify = ">=3.0.2"
python-stdnum = "^1.17"
python-swiftclient = "3.9.0"
python-u2flib-server = "^5.0.1"
python-u2flib_server = "^5.0.1"
pytz = "2017.3"
pyusb = "1.0.0"
pyzbar = "^0.1.8"
qrcode = "5.3"
raven = "^6.10.0"
redis = "2.10.5"
reportlab = "3.3.0"
requests = "2.20.0"
requests-mock = "^1.9.3"
requests-oauthlib = "1.1.0"
requests-toolbelt = "0.9.1"
retrying = "^1.3.3"
serial = "^0.0.97"
shapely = "1.6.4.post2"
simplejson = "^3.17.6"
slugify = "^0.0.1"
sphinx = ">=1.6.7"
sphinx-intl = "^2.0.1"
sphinx-patchqueue = ">=1.0"
sphinx_rtd_theme = "^1.0.0"
sqlalchemy = "^1.4.31"
statsd = "3.2.1"
suds-jurko = "0.6"
toml = "^0.10.2"
twython = "^3.9.1"
unidecode = "1.0.22"
unidiff = "^0.7.0"
vatnumber = "1.2"
vcrpy = ">=2.1.1"
vcrpy-unittest = "^0.1.7"
vobject = "0.9.3"
voicent-python = "^1.0"
webcolors = "1.10"
websocket-client = "^1.2.3"
wechatpy = "^1.8.18"
wget = "^3.2"
wheel = "^0.37.1"
xlrd = "1.0.0"
xlwt = "1.3"
xmlformatter = "^0.2.4"
xmltodict = "^0.12.0"
zeep = "^4.1.0"
zxcvbn = "^4.4.28"

[tool.poetry.dev-dependencies]
websocket_client = "^1.1.0"

[tool.poetry.dependencies.black]
git = "https://github.com/psf/black.git"
rev = "21.6b0"

[tool.poetry.dependencies.pylint-odoo]
git = "https://github.com/oca/pylint-odoo.git"
